# services/ml/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI Leader ID ML API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/internal/v1/transcribe": {
      "post": {
        "summary": "Transcribe",
        "operationId": "transcribe_internal_v1_transcribe_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TranscribeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TranscribeResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/internal/v1/speech": {
      "post": {
        "summary": "Speech",
        "operationId": "speech_internal_v1_speech_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SpeechRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "audio/mpeg": {}
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/internal/v1/consistency": {
      "post": {
        "summary": "Consistency",
        "operationId": "consistency_internal_v1_consistency_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsistencyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsistencyResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/internal/v1/surprise-question": {
      "post": {
        "summary": "Surprise Question",
        "operationId": "surprise_question_internal_v1_surprise_question_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SurpriseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SurpriseResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/internal/v1/usage": {
      "get": {
        "summary": "Usage",
        "operationId": "usage_internal_v1_usage_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Usage"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/internal/v1/interview/follow-up": {
      "post": {
        "summary": "Interview Follow Up",
        "operationId": "interview_follow_up_internal_v1_interview_follow_up_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FollowUpRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FollowUpResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/internal/v1/interview/draft": {
      "post": {
        "summary": "Interview Draft",
        "operationId": "interview_draft_internal_v1_interview_draft_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DraftRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DraftResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/internal/v1/quality-check": {
      "post": {
        "summary": "Quality Check",
        "operationId": "quality_check_internal_v1_quality_check_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QualityCheckRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QualityCheckResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "CandidateView": {
        "properties": {
          "candidateId": {
            "type": "string",
            "title": "Candidateid"
          },
          "application": {
            "$ref": "#/components/schemas/Answers"
          },
          "test": {
            "$ref": "#/components/schemas/TestAnswers"
          },
          "englishCertificate": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EnglishCertificate"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "candidateId",
          "application",
          "test"
        ],
        "title": "CandidateView"
      },
...
      "ConsistencyItem": {
        "properties": {
          "itemId": {
            "type": "string",
            "title": "Itemid"
          },
          "topic": {
            "type": "string",
            "enum": [
              "english",
              "invision_knowledge",
              "motivation",
              "experience",
              "achievements",
              "other"
            ],
            "title": "Topic"
          },
          "claim": {
            "$ref": "#/components/schemas/Claim"
          },
          "observation": {
            "$ref": "#/components/schemas/Observation"
          },
          "status": {
            "type": "string",
            "enum": [
              "consistent",
              "discrepancy",
              "unverified",
              "confirmed",
              "resolved"
            ],
            "title": "Status"
          },
          "whatToDo": {
            "type": "string",
            "title": "Whattodo"
          },
          "askInInterview": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Askininterview"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "itemId",
          "topic",
          "claim",
          "observation",
          "status",
          "whatToDo",
          "askInInterview"
        ],
        "title": "ConsistencyItem"
      },
      "ConsistencyRequest": {
        "properties": {
          "stage": {
            "type": "string",
            "enum": [
              "before",
              "after"
            ],
            "title": "Stage"
          },
          "candidate": {
            "$ref": "#/components/schemas/CandidateView"
          },
          "simulationEnglish": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EnglishMetrics"
              },
              {
                "type": "null"
              }
            ]
          },
          "simulationTurns": {
            "items": {
              "$ref": "#/components/schemas/Turn"
            },
            "type": "array",
            "title": "Simulationturns",
            "default": []
          },
          "interviewTranscript": {
            "items": {
              "$ref": "#/components/schemas/InterviewTurn"
            },
            "type": "array",
            "title": "Interviewtranscript",
            "default": []
          },
          "beforeItems": {
            "items": {
              "$ref": "#/components/schemas/ConsistencyItem"
            },
            "type": "array",
            "title": "Beforeitems"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "stage",
          "candidate"
        ],
        "title": "ConsistencyRequest"
      },
      "ConsistencyResult": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ConsistencyItem"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items"
        ],
        "title": "ConsistencyResult"
      },
...
      "DraftRequest": {
        "properties": {
          "candidateId": {
            "type": "string",
            "title": "Candidateid"
          },
          "transcript": {
            "items": {
              "$ref": "#/components/schemas/InterviewTurn"
            },
            "type": "array",
            "title": "Transcript"
          },
          "notes": {
            "items": {
              "$ref": "#/components/schemas/InterviewNote"
            },
            "type": "array",
            "title": "Notes",
            "default": []
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "candidateId",
          "transcript"
        ],
        "title": "DraftRequest",
        "description": "The transcript and any notes. The interviewer's scores are deliberately absent, and extra fields are refused."
      },
      "DraftResult": {
        "properties": {
          "scores": {
            "items": {
              "$ref": "#/components/schemas/DriveScore"
            },
            "type": "array",
            "title": "Scores"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "scores"
        ],
        "title": "DraftResult"
      },
...
      "EarlierSuggestion": {
        "properties": {
          "question": {
            "type": "string",
            "title": "Question"
          },
          "competency": {
            "type": "string",
            "enum": [
              "D",
              "R",
              "I",
              "V",
              "E"
            ],
            "title": "Competency"
          },
          "status": {
            "type": "string",
            "enum": [
              "open",
              "asked",
              "dismissed"
            ],
            "title": "Status"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "question",
          "competency",
          "status"
        ],
        "title": "EarlierSuggestion"
      },
...
      "EnglishMetrics": {
        "properties": {
          "wordsPerMinute": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wordsperminute"
          },
          "fillerRate": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fillerrate"
          },
          "meanTurnLength": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Meanturnlength"
          },
          "lexicalDiversity": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lexicaldiversity"
          },
          "grammarErrorsPer100Words": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grammarerrorsper100Words"
          },
          "cefrEstimate": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cefrestimate"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "wordsPerMinute",
          "fillerRate",
          "meanTurnLength",
          "lexicalDiversity",
          "grammarErrorsPer100Words",
          "cefrEstimate"
        ],
        "title": "EnglishMetrics"
      },
...
      "FollowUpRequest": {
        "properties": {
          "candidate": {
            "$ref": "#/components/schemas/CandidateView"
          },
          "plannedQuestions": {
            "items": {
              "$ref": "#/components/schemas/PlannedQuestion"
            },
            "type": "array",
            "title": "Plannedquestions",
            "default": []
          },
          "question": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Question"
          },
          "answer": {
            "items": {
              "$ref": "#/components/schemas/FollowUpSegment"
            },
            "type": "array",
            "minItems": 1,
            "title": "Answer"
          },
          "earlier": {
            "items": {
              "$ref": "#/components/schemas/EarlierSuggestion"
            },
            "type": "array",
            "title": "Earlier",
            "default": []
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "candidate",
          "answer"
        ],
        "title": "FollowUpRequest"
      },
      "FollowUpResult": {
        "properties": {
          "suggestions": {
            "items": {
              "$ref": "#/components/schemas/FollowUpSuggestion"
            },
            "type": "array",
            "maxItems": 2,
            "title": "Suggestions"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "suggestions"
        ],
        "title": "FollowUpResult"
      },
      "FollowUpSegment": {
        "properties": {
          "segmentId": {
            "type": "string",
            "pattern": "^fseg_\\d{2,}$",
            "title": "Segmentid"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "startSec": {
            "type": "number",
            "title": "Startsec"
          },
          "endSec": {
            "type": "number",
            "title": "Endsec"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "segmentId",
          "text",
          "startSec",
          "endSec"
        ],
        "title": "FollowUpSegment"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "InterviewNote": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "text": {
            "type": "string",
            "title": "Text"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "text"
        ],
        "title": "InterviewNote"
      },
...
      "InterviewTurn": {
        "properties": {
          "turnId": {
            "type": "string",
            "pattern": "^iturn_\\d{2,}$",
            "title": "Turnid"
          },
          "speaker": {
            "type": "string",
            "enum": [
              "interviewer",
              "candidate"
            ],
            "title": "Speaker"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "startSec": {
            "type": "number",
            "title": "Startsec"
          },
          "endSec": {
            "type": "number",
            "title": "Endsec"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "turnId",
          "speaker",
          "text",
          "startSec",
          "endSec"
        ],
        "title": "InterviewTurn"
      },
...
      "PlannedQuestion": {
        "properties": {
          "focus": {
            "type": "string",
            "title": "Focus"
          },
          "question": {
            "type": "string",
            "title": "Question"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "focus",
          "question"
        ],
        "title": "PlannedQuestion"
      },
      "QualityCheckRequest": {
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "interview",
              "calibration"
            ],
            "title": "Kind"
          },
          "transcript": {
            "items": {
              "$ref": "#/components/schemas/InterviewTurn"
            },
            "type": "array",
            "title": "Transcript",
            "default": []
          },
          "history": {
            "items": {
              "$ref": "#/components/schemas/ScoredInterview"
            },
            "type": "array",
            "title": "History",
            "default": []
          },
          "interviewerRef": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Interviewerref"
          },
          "periodFrom": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Periodfrom"
          },
          "periodTo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Periodto"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "kind"
        ],
        "title": "QualityCheckRequest"
      },
      "QualityCheckResult": {
        "properties": {
          "signals": {
            "items": {
              "$ref": "#/components/schemas/QualitySignal"
            },
            "type": "array",
            "title": "Signals"
          },
          "talkShare": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TalkShare"
              },
              {
                "type": "null"
              }
            ]
          },
          "drift": {
            "items": {
              "$ref": "#/components/schemas/Drift"
            },
            "type": "array",
            "title": "Drift",
            "default": []
          },
          "interviews": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Interviews"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "signals"
        ],
        "title": "QualityCheckResult"
      },
...
      "SpeechRequest": {
        "properties": {
          "text": {
            "type": "string",
            "title": "Text"
          },
          "scenarioId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scenarioid"
          },
          "voice": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Voice"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "text"
        ],
        "title": "SpeechRequest"
      },
      "SurpriseRequest": {
        "properties": {
          "candidate": {
            "$ref": "#/components/schemas/CandidateView"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "candidate"
        ],
        "title": "SurpriseRequest"
      },
      "SurpriseResult": {
        "properties": {
          "question": {
            "type": "string",
            "title": "Question"
          },
          "competency": {
            "type": "string",
            "enum": [
              "D",
              "R",
              "I",
              "V",
              "E"
            ],
            "title": "Competency"
          },
          "why": {
            "type": "string",
            "title": "Why"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "question",
          "competency",
          "why"
        ],
        "title": "SurpriseResult"
      },
...
      "TranscribeRequest": {
        "properties": {
          "purpose": {
            "type": "string",
            "enum": [
              "interview",
              "surprise",
              "turn",
              "presentation",
              "follow_up"
            ],
            "title": "Purpose"
          },
          "audioRef": {
            "type": "string",
            "title": "Audioref"
          },
          "language": {
            "type": "string",
            "const": "en",
            "title": "Language",
            "default": "en"
          },
          "speakers": {
            "type": "integer",
            "enum": [
              1,
              2
            ],
            "title": "Speakers",
            "default": 2
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "purpose",
          "audioRef"
        ],
        "title": "TranscribeRequest"
      },
      "TranscribeResult": {
        "properties": {
          "turns": {
            "items": {
              "$ref": "#/components/schemas/TranscribedTurn"
            },
            "type": "array",
            "title": "Turns"
          },
          "durationSec": {
            "type": "number",
            "title": "Durationsec"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "turns",
          "durationSec"
        ],
        "title": "TranscribeResult"
      },
      "TranscribedTurn": {
        "properties": {
          "speaker": {
            "type": "string",
            "enum": [
              "interviewer",
              "candidate"
            ],
            "title": "Speaker"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "startSec": {
            "type": "number",
            "title": "Startsec"
          },
          "endSec": {
            "type": "number",
            "title": "Endsec"
          },
          "confidence": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "speaker",
          "text",
          "startSec",
          "endSec"
        ],
        "title": "TranscribedTurn"
      },
      "Turn": {
        "properties": {
          "turnId": {
            "type": "string",
            "pattern": "^turn_\\d{2,}$",
            "title": "Turnid"
          },
          "speaker": {
            "type": "string",
            "enum": [
              "candidate",
              "character"
            ],
            "title": "Speaker"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "startedAt": {
            "type": "string",
            "title": "Startedat"
          },
          "endedAt": {
            "type": "string",
            "title": "Endedat"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "turnId",
          "speaker",
          "text",
          "startedAt",
          "endedAt"
        ],
        "title": "Turn"
      },
...
      "Usage": {
        "properties": {
          "gatewayMode": {
            "type": "string",
            "enum": [
              "live",
              "record",
              "replay"
            ],
            "title": "Gatewaymode"
          },
          "liveCalls": {
            "type": "integer",
            "title": "Livecalls"
          },
          "replayedCalls": {
            "type": "integer",
            "title": "Replayedcalls"
          },
          "spentUsd": {
            "type": "number",
            "title": "Spentusd"
          },
          "capUsd": {
            "type": "number",
            "title": "Capusd"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "gatewayMode",
          "liveCalls",
          "replayedCalls",
          "spentUsd",
          "capUsd"
        ],
        "title": "Usage"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Internal-Token"
      }
    }
  }
}